back-end/requirements.txt
fastapi==0.104.0
pycryptodome==3.19.0
pydantic==1.10.4
PyJWT==2.8.0
SQLAlchemy==2.0.22
